This is the test demo for "Deep networks for compressed image sensing." (ICME2017)
%------------------------------------------------------------------------------------------------------
%This is the test demo for "Deep networks for compressed image sensing."
(ICME2017)
%Author: Wuzhen Shi
%Email: wzhshi@hit.edu.cn
%School of Computer Science of Technology, Harbin Institute of
Technology

%-------------------------------------------------------------------------------------------------------

%BibTex:
% @inproceedings{Shi2017Deep,
%   title={Deep networks for compressed image sensing},
%   author={Shi, Wuzhen and Jiang, Feng and Zhang, Shengping and Zhao,
Debin},
%   booktitle={IEEE International Conference on Multimedia and Expo},
%   pages={877-882},
%   year={2017},
% }

%Usage:
%set different samplingRate (0.1 to 0.5) to obtain the same results as
we report in the ICME2017 paper.

Two natural questions to compressed sensing are how to design an
effective sampling matrix, and how to develop a fast nonlinear
reconstruction algorithm. They are very challenging. In this paper, we
also focus on how to use deep networks to deal with this two problems.
To the problem how to design an effective sampling matrix, we learn the
sampling matrix by a convolutional neural network automatically. Given
the CS measurements, our method use a deep network to recover the
original image signal.